Add task barrier to workflow graph
Add a generic barrier policy to workflow graph. Move Mistral specific join to the Mistral workflow graph composer.
