+++ orchestra/tests/unit/composition/mistral/test_workflow_graph.py
@@ -32,7 +32,7 @@ EXPECTED_WF_GRAPH = {
         },
         {
             'id': 'task5',
-            'join': True
+            'barrier': '*'
         },
         {
             'id': 'task6'
@@ -138,12 +138,21 @@ class WorkflowGraphTest(base.WorkflowGraphTest):
         wf_graph.add_sequence('task8', 'task9')
 
     def _update_tasks_attrs(self, wf_graph):
-        wf_graph.update_task('task5', join=True)
+        wf_graph.update_task('task5', barrier='*')
 
     def _prep_graph(self, wf_graph):
         self._add_tasks(wf_graph)
         self._update_tasks_attrs(wf_graph)
         self._add_sequences(wf_graph)
+
+    def _is_join_task(self, wf_graph, task_name):
+        return wf_graph.has_barrier(task_name)
+
+    def _is_split_task(self, wf_graph, task_name):
+        return (
+            len(wf_graph.get_prev_sequences(task_name)) > 1 and
+            not self._is_join_task(wf_graph, task_name)
+        )
 
     def test_basic_graph(self):
         wf_graph = composition.WorkflowGraph()
@@ -214,12 +223,12 @@ class WorkflowGraphTest(base.WorkflowGraphTest):
         wf_graph = composition.WorkflowGraph()
         self._prep_graph(wf_graph)
 
-        self.assertTrue(wf_graph.is_join_task('task5'))
-        self.assertFalse(wf_graph.is_join_task('task9'))
+        self.assertTrue(self._is_join_task(wf_graph, 'task5'))
+        self.assertFalse(self._is_join_task(wf_graph, 'task9'))
 
     def test_is_split_task(self):
         wf_graph = composition.WorkflowGraph()
         self._prep_graph(wf_graph)
 
-        self.assertFalse(wf_graph.is_split_task('task5'))
-        self.assertTrue(wf_graph.is_split_task('task9'))
+        self.assertFalse(self._is_split_task(wf_graph, 'task5'))
+        self.assertTrue(self._is_split_task(wf_graph, 'task9'))
